caseLog/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.3
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
filelock==3.6.0
identify==2.4.11
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
loadtest==0.0.2
MarkupSafe==2.1.0
nodeenv==1.6.0
packaging==21.3
platformdirs==2.5.1
pre-commit==2.17.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.3
wrapt==1.14.0
